requirements.txt
# Core
Django==2.2.6
djangorestframework==3.10.3
django-filter==2.2.0
Markdown==3.0.1
djangorestframework_simplejwt
django-configurations

# For persistense stores
psycopg2-binary==2.8.4

# Dev tools
autopep8==1.4.4
flake8==3.7.8
pycodestyle==2.5.0
pyflakes==2.1.1
sentry-sdk==0.10.2

# Log
python-json-logger==0.1.11

# Testing
mock==3.0.5
django-nose==1.4.6
nose==1.3.7
nose-progressive==1.5.2
coverage==4.5.4
factory-boy==2.11.1

# Static storage
django-storages==1.7.2

# Other dependencies
django-redis-cache
pyyaml==5.1.2
djangorestframework-camel-case==1.1.2
django-cors-headers==3.2.0
validate_email
